Rename mock_ridsp to mock_uss
