+++ monitoring/README.md
@@ -67,7 +67,9 @@ OAuth server](../cmds/dummy-oauth) which grants properly-formatted access tokens
 (which can be validated against the
 [test public key](../build/test-certs/auth2.pem)) to anyone requesting them.
 
-### mock_ridsp
+### mock_uss
+
+#### ridsp
 
 ![mock_ridsp system](../assets/mock_ridsp_system.png)
