+++ monitoring/mock_uss/Dockerfile
@@ -3,10 +3,10 @@ FROM python:3.8
 # `docker build` should be run from `monitoring` (the parent folder of this folder)
 RUN apt-get update && apt-get install openssl && apt-get install ca-certificates
 RUN mkdir -p /app/monitoring/monitorlib
-RUN mkdir /app/monitoring/mock_ridsp
+RUN mkdir /app/monitoring/mock_uss
 COPY monitorlib/requirements.txt /app/monitoring/monitorlib/requirements.txt
-COPY mock_ridsp/requirements.txt /app/monitoring/mock_ridsp/requirements.txt
-WORKDIR /app/monitoring/mock_ridsp
+COPY mock_uss/requirements.txt /app/monitoring/mock_uss/requirements.txt
+WORKDIR /app/monitoring/mock_uss
 RUN pip install -r requirements.txt
 RUN rm -rf __pycache__
 ADD . /app/monitoring
