+++ monitoring/mock_uss/README.md
@@ -1,0 +1,4 @@
+`mock_uss` contains mock implementations of a number of USS services.  It
+provides a development-level web server that responds to requests to the USS
+endpoints defined in the relevant ASTM standards in a standards-compliant
+manner.
