+++ monitoring/mock_uss/config.py
@@ -3,11 +3,12 @@ import os
 from monitoring.monitorlib import auth_validation
 
 
-ENV_KEY_PREFIX = 'MOCK_RIDSP'
+ENV_KEY_PREFIX = 'MOCK_USS'
 ENV_KEY_PUBLIC_KEY = '{}_PUBLIC_KEY'.format(ENV_KEY_PREFIX)
 ENV_KEY_TOKEN_AUDIENCE = '{}_TOKEN_AUDIENCE'.format(ENV_KEY_PREFIX)
 ENV_KEY_BASE_URL = '{}_BASE_URL'.format(ENV_KEY_PREFIX)
 ENV_KEY_AUTH = '{}_AUTH_SPEC'.format(ENV_KEY_PREFIX)
+ENV_KEY_SERVICES = '{}_SERVICES'.format(ENV_KEY_PREFIX)
 ENV_KEY_DSS = '{}_DSS_URL'.format(ENV_KEY_PREFIX)
 
 # These keys map to entries in the Config class
@@ -15,6 +16,7 @@ KEY_TOKEN_PUBLIC_KEY = 'TOKEN_PUBLIC_KEY'
 KEY_TOKEN_AUDIENCE = 'TOKEN_AUDIENCE'
 KEY_BASE_URL = 'USS_BASE_URL'
 KEY_AUTH_SPEC = 'AUTH_SPEC'
+KEY_SERVICES = 'SERVICES'
 KEY_DSS_URL = 'DSS_URL'
 
 
@@ -27,4 +29,5 @@ class Config(object):
   TOKEN_AUDIENCE = os.environ.get(ENV_KEY_TOKEN_AUDIENCE, '')
   USS_BASE_URL = os.environ[ENV_KEY_BASE_URL]
   AUTH_SPEC = os.environ[ENV_KEY_AUTH]
+  SERVICES = set(svc.strip().lower() for svc in os.environ.get(ENV_KEY_SERVICES, '').split(','))
   DSS_URL = os.environ[ENV_KEY_DSS]
