+++ monitoring/mock_uss/ridsp/routes.py
@@ -1,0 +1,10 @@
+from monitoring.mock_uss import webapp
+
+
+@webapp.route('/ridsp/status')
+def ridsp_status():
+  return 'Mock RID Service Provider ok'
+
+
+from . import routes_ridsp
+from . import routes_injection
